Remove kicker line sensor bouncy, and some extra

+++ src/subsystemfile/intake.cpp
@@ -8,7 +8,7 @@ void setIntake() {
     bool L1TON = controller.get_digital(pros::E_CONTROLLER_DIGITAL_L1);
     bool L2TON = controller.get_digital(pros::E_CONTROLLER_DIGITAL_L2);
 
-    float intakePower = 127 * (L1TON - L2TON);
-
+    int intakePower = 127 * (L1TON - L2TON);
+    
     setIntakeVoltage(intakePower);
 }
